+++ tests/test_hamiltonians/test_save.py
@@ -18,7 +18,7 @@ def test_save_graphene(tmp_path: Path) -> None:
             lattice_constant=np.sqrt(3),
             chemical_potential=-1,
             hubbard_int_orbital_basis=[1.0, 1.0],
-            delta=np.ones(2, dtype=np.complex64),
+            delta=np.ones(2, dtype=np.complex128),
         )
     )
     file_path = tmp_path / "test.hdf5"
@@ -39,7 +39,7 @@ def test_save_graphene_with_beta_and_q(tmp_path: Path) -> None:
             hubbard_int_orbital_basis=[1.0, 1.0],
             q=np.ones(2, dtype=np.float64),
             beta=100,
-            delta=np.ones(2, dtype=np.complex64),
+            delta=np.ones(2, dtype=np.complex128),
         )
     )
     file_path = tmp_path / "test.hdf5"
@@ -60,7 +60,7 @@ def test_save_dressed_graphene(tmp_path: Path) -> None:
             lattice_constant=np.sqrt(3),
             chemical_potential=0,
             hubbard_int_orbital_basis=[1.0, 1.0, 1.0],
-            delta=np.ones(3, dtype=np.complex64),
+            delta=np.ones(3, dtype=np.complex128),
         )
     )
     file_path = tmp_path / "test.hdf5"
@@ -83,7 +83,7 @@ def test_save_dressed_graphene_with_q_and_beta(tmp_path: Path) -> None:
             hubbard_int_orbital_basis=[1.0, 1.0, 1.0],
             q=np.ones(2, dtype=np.float64),
             beta=100,
-            delta=np.ones(3, dtype=np.complex64),
+            delta=np.ones(3, dtype=np.complex128),
         )
     )
     file_path = tmp_path / "test.hdf5"
@@ -102,7 +102,7 @@ def test_save_one_band(tmp_path: Path) -> None:
             lattice_constant=1,
             chemical_potential=0,
             hubbard_int_orbital_basis=[1.0],
-            delta=np.ones(1, dtype=np.complex64),
+            delta=np.ones(1, dtype=np.complex128),
         )
     )
     file_path = tmp_path / "test.hdf5"
@@ -123,7 +123,7 @@ def test_save_one_band_with_q_and_beta(tmp_path: Path) -> None:
             hubbard_int_orbital_basis=[1.0],
             q=np.ones(2, dtype=np.float64),
             beta=100,
-            delta=np.ones(1, dtype=np.complex64),
+            delta=np.ones(1, dtype=np.complex128),
         )
     )
     file_path = tmp_path / "test.hdf5"
@@ -142,7 +142,7 @@ def test_save_two_band(tmp_path: Path) -> None:
             lattice_constant=1,
             chemical_potential=0,
             hubbard_int_orbital_basis=[1.0, 1.0],
-            delta=np.ones(2, dtype=np.complex64),
+            delta=np.ones(2, dtype=np.complex128),
         )
     )
     file_path = tmp_path / "test.hdf5"
@@ -181,7 +181,7 @@ def test_save_three_band(tmp_path: Path) -> None:
             lattice_constant=1,
             chemical_potential=0,
             hubbard_int_orbital_basis=[1.0, 1.0, 1.0],
-            delta=np.ones(3, dtype=np.complex64),
+            delta=np.ones(3, dtype=np.complex128),
         )
     )
     file_path = tmp_path / "test.hdf5"
